SavePoint

+++ app/headers/scene.h
@@ -23,7 +23,7 @@ public:
     PointCloud &getPointCloudAtIndex(const int &index);
 
     void generateImport(const std::string &filename, const float &decimate);
-    void generatePrimitive(const PopupType &type, const double &resolution);
+    void generatePrimitive(const PopupType &type, const int &resolution);
 };
 
 #endif
